support for various --virt- options for both profile and system, add --power-type to system
We add various --virt options to both profile, and system. We complement
profile by adding a --comment option as it is now supported by all
components of cobbler.
Next we add the following --virt options:

    --virt-auto-boot=VIRT_AUTO_BOOT
        --virt-cpus=VIRT_CPUS
        --virt-file-size=VIRT_FILE_SIZE
        --virt-disk-driver=VIRT_DISK_DRIVER
        --virt-ram=VIRT_RAM
        --virt-type=VIRT_TYPE
        --virt-path=VIRT_PATH

The following option is only add to profile, since in system it's part
of --interfaces:

        --virt-bridge=VIRT_BRIDGE

Finally we also add --power-type to system. To support this option, you
have to install the fence-agents package.

The code is now more streamlined in all modules: We have removed every instance
of defaultto that set the value to '', because this caused puppet to call

     cobbler <subsystem> edit --name=:name --<option>=''