new options: virt-*, power-type, kernel-options...
and many more!

we add all major virt-options to both, profile and system.
virt-bridge we only explicitly add to profile, providing a default for
systems that use this profile.
Otherwise, system's interfaces already handles virt-bridge!

We also add power-type to system. In order to make use of it, the
fence-agents package is needed.

Finally, we also add the nameservers option to profile, and also add a
nameserver search option.