More sane DHCP router setup
Previously the DHCP router address was constructed for each interface by taking
the first three octets from the IPv4 address of the interface, and appending
".1" to the end. This broke cobblersync if that particular .1 address did not
exist. This commit allows overriding the router advertised by dhcpd with the
$dhcp_router parameter - however due to current limitations of this module the
same $dhcp_router is used for all interfaces, _if_ the parameter is defined. If
the parameter is not defined, then the router address is set to the IPv4 address
of that particular interface.

Signed-off-by: Samuli Seppänen <samuli.seppanen@gmail.com>